organize components into separate files in src
